Intelli Hire AI-Powered Resume Screening System/requirements.txt
flask==3.0.0
google-generativeai==0.3.2
chromadb==0.4.22
sentence-transformers==2.3.1
PyMuPDF==1.23.8
python-docx==1.1.0
supabase==2.3.0
python-dotenv==1.0.0
numpy==1.24.3